chore: increase threshold spammer

+++ crates/topos-certificate-spammer/src/lib.rs
@@ -38,7 +38,7 @@ pub struct CertificateSpammerConfig {
     pub target_nodes: Option<Vec<String>>,
     pub target_nodes_path: Option<String>,
     pub local_key_seed: u64,
-    pub cert_per_batch: u8,
+    pub cert_per_batch: u64,
     pub nb_subnets: u8,
     pub nb_batches: Option<u64>,
     pub batch_interval: u64,
